requirements.txt
Django
django-countries
django-localeurl
Pillow
requests
django-language-field

# Broadcast analysis
inaSpeechSegmenter # CNN-based speech/music/noise segmentation for broadcast audio
# webrtcvad        # Voice Activity Detection — replaced by inaSpeechSegmenter (kept for rollback)
faster-whisper     # Local speech transcription (CPU or GPU)
pyacoustid         # AcoustID song identification (needs fpcalc binary)
openai             # API transcription fallback + LLM summarisation
#anthropic          # Claude API for audio transcription + LLM provider
ollama             # Local LLM summarisation via Ollama

# REST API
djangorestframework       # DRF — REST API framework
django-filter             # Structured filtering for DRF
djangorestframework-simplejwt  # JWT authentication for API clients
